Attempt to reduce glitching by avoiding simultaneous updates
I've been chasing a scrolling glitch that I'm fairly confident is eliminated
by not enabling serial (see followup commit). But in the process of
troubleshooting it, I found that it only seems to happen when the scroll text
and clock are updated on the same frame. While I don't yet understand why
that's related, I prefer a belt and suspenders approach here.

The only downside is that if the scrolling speed is set so high that it
happens on every frame, there will be no opportunity to update the clock. That
can be easily avoided by reducing the speed.