Implement a basic Conway's Life module
The hard-code-and-recompile nature of this project is turning into a
bit of tech debt now that I want to have multiple marquees with
different configurations. I'd also like to be able to change them
on the fly.

This was a quick Sunday afternoon hack and is not very clean or
extensible, so it's going in a branch. I'm considering options
for a more flexible redesign. In the meantime, though, it does
look pretty cool: https://youtu.be/i_uLQUjtJxU